+++ discord/eventdelegation.go
@@ -2,41 +2,12 @@ package discord
 
 import (
 	"fmt"
-	"io"
 	"time"
 
 	"github.com/andersfylling/disgord"
-	"github.com/jonas747/dca"
 )
 
-// TO-DO All of the voice related logic will be moved to the yt package
-// Getting everything working in here first
-
-// VoiceChannelCache keeps a map record of userID's and the channelID that user currently belongs to
-type VoiceChannelCache struct {
-	Cache map[disgord.Snowflake]disgord.Snowflake
-}
-
-// NewVoiceChannelCache return a pointer to a new voicechannelcache
-func NewVoiceChannelCache() *VoiceChannelCache {
-	cacheMap := make(map[disgord.Snowflake]disgord.Snowflake)
-	return &VoiceChannelCache{
-		Cache: cacheMap,
-	}
-}
-
-var vcCache = NewVoiceChannelCache()
-
-// UpdateCache updates the voicechannel cache based upon the set channel id on voice state updates
-// A channelID of 0 means a user left, in that case remove them from the cache
-func (vcc *VoiceChannelCache) UpdateCache(chID disgord.Snowflake, uID disgord.Snowflake) {
-	switch chID {
-	case 0:
-		delete(vcCache.Cache, uID)
-	default:
-		vcCache.Cache[uID] = chID
-	}
-}
+var vcCache = NewVoiceChannelCache(disgordGlobalClient)
 
 // Using this for access to the global clients FOR NOW as passing it through the handlers has proven tricky
 // TO-DO find a solution to get rid of the global variables, including the client
@@ -53,7 +24,8 @@ func RespondToCommand(s disgord.Session, data *disgord.MessageCreate) {
 	fmt.Printf("Command %+v by user %+v | %+v\n", command, user.Username, time.Now().Format("Mon Jan _2 15:04:05 2006"))
 	switch command {
 	case "play":
-		go processAndPlay(s, data, args[0])
+		go ProcessAndPlay(data.Message.GuildID, data.Message.Author.ID, args[0], disgordGlobalClient)
+		go deleteMessage(data.Message, 20*time.Second, disgordGlobalClient)
 	default:
 		cec.Delegate()
 	}
@@ -100,91 +72,91 @@ func RespondToVoiceChannelUpdate(s disgord.Session, data *disgord.VoiceStateUpda
 	vcCache.UpdateCache(data.ChannelID, data.UserID)
 }
 
-func processAndPlay(s disgord.Session, data *disgord.MessageCreate, arg string) {
-	// ss := youtube.NewSearchService(ytService)
-	// ytc := yt.NewYoutubeClient(ss)
-	// _, err := ytc.SearchAndDownload(arg)
-	// if err != nil {
-	// 	fmt.Printf("\nERROR WITH FILE: %+v\n", err)
-	// }
+// func processAndPlay(s disgord.Session, data *disgord.MessageCreate, arg string) {
+// 	// ss := youtube.NewSearchService(ytService)
+// 	// ytc := yt.NewYoutubeClient(ss)
+// 	// _, err := ytc.SearchAndDownload(arg)
+// 	// if err != nil {
+// 	// 	fmt.Printf("\nERROR WITH FILE: %+v\n", err)
+// 	// }
 
-	go deleteMessage(data.Message, 20*time.Second, disgordGlobalClient)
-	requestURL := fmt.Sprintf("http://localhost:8080/mp3/%+v", arg)
-	encodeSess, err := dca.EncodeFile(requestURL, &dca.EncodeOptions{
-		Volume:           256,
-		Channels:         2,
-		FrameRate:        48000,
-		FrameDuration:    20,
-		Bitrate:          64,
-		Application:      "audio",
-		CompressionLevel: 5,
-		PacketLoss:       1,
-		BufferedFrames:   200, // At 20ms frames that's 2s
-		VBR:              true,
-		StartTime:        0,
-		RawOutput:        true,
-		Threads:          0,
-	})
-	if err != nil {
-		fmt.Printf("\nERROR ENCODING: %+v\n", err)
-	}
+// 	go deleteMessage(data.Message, 20*time.Second, disgordGlobalClient)
+// 	requestURL := fmt.Sprintf("http://localhost:8080/mp3/%+v", arg)
+// 	encodeSess, err := dca.EncodeFile(requestURL, &dca.EncodeOptions{
+// 		Volume:           256,
+// 		Channels:         2,
+// 		FrameRate:        48000,
+// 		FrameDuration:    20,
+// 		Bitrate:          64,
+// 		Application:      "audio",
+// 		CompressionLevel: 5,
+// 		PacketLoss:       1,
+// 		BufferedFrames:   200, // At 20ms frames that's 2s
+// 		VBR:              true,
+// 		StartTime:        0,
+// 		RawOutput:        true,
+// 		Threads:          0,
+// 	})
+// 	if err != nil {
+// 		fmt.Printf("\nERROR ENCODING: %+v\n", err)
+// 	}
 
-	// fmt.Printf("\nENCODESESS: %+v", encodeSess)
+// 	// fmt.Printf("\nENCODESESS: %+v", encodeSess)
 
-	// fmt.Printf("\nGUILDID: %+v | CHANNELID: %+v\n", data.Message.GuildID, data.Message.ChannelID)
+// 	// fmt.Printf("\nGUILDID: %+v | CHANNELID: %+v\n", data.Message.GuildID, data.Message.ChannelID)
 
-	vc, err := disgordGlobalClient.VoiceConnectOptions(data.Message.GuildID, vcCache.Cache[data.Message.Author.ID], true, false)
-	if err != nil {
-		fmt.Printf("\nERROR CONNECTING TO VOICE CHANNEL: %+v\n", err)
-		// return
-	}
-	err = vc.StartSpeaking()
-	if err != nil {
-		fmt.Printf("\nERROR SPEAKING: %+v\n", err)
-	}
+// 	vc, err := disgordGlobalClient.VoiceConnectOptions(data.Message.GuildID, vcCache.Cache[data.Message.Author.ID], true, false)
+// 	if err != nil {
+// 		fmt.Printf("\nERROR CONNECTING TO VOICE CHANNEL: %+v\n", err)
+// 		// return
+// 	}
+// 	err = vc.StartSpeaking()
+// 	if err != nil {
+// 		fmt.Printf("\nERROR SPEAKING: %+v\n", err)
+// 	}
 
-	ticker := time.NewTicker(20 * time.Millisecond)
-	done := make(chan bool)
-	eofChannel := make(chan bool)
+// 	ticker := time.NewTicker(20 * time.Millisecond)
+// 	done := make(chan bool)
+// 	eofChannel := make(chan bool)
 
-	go func() {
-		defer encodeSess.Cleanup()
-		for {
-			select {
-			case <-done:
-				err := vc.StopSpeaking()
-				if err != nil && err != io.EOF {
-					fmt.Printf("\nERROR STOPPING TALKING: %+v\n", err)
-				}
-				time.Sleep(1 * time.Second)
-				err = vc.Close()
-				if err != nil && err != io.EOF {
-					fmt.Printf("\nERROR LEAVING VC: %+v\n", err)
-				}
-				fmt.Println("\nLEAVING VOICE CHANNEL")
-				return
-			case <-ticker.C:
-				nextFrame, err := encodeSess.OpusFrame()
-				if err != nil && err != io.EOF {
-					fmt.Printf("\nERROR PLAYING DCA: %+v\n", err)
-				}
-				if err == io.EOF {
-					fmt.Println("\nPLAYBACK FINISHED")
-					eofChannel <- true
-				}
-				err = vc.SendOpusFrame(nextFrame)
-			}
-		}
-	}()
+// 	go func() {
+// 		defer encodeSess.Cleanup()
+// 		for {
+// 			select {
+// 			case <-done:
+// 				err := vc.StopSpeaking()
+// 				if err != nil && err != io.EOF {
+// 					fmt.Printf("\nERROR STOPPING TALKING: %+v\n", err)
+// 				}
+// 				time.Sleep(1 * time.Second)
+// 				err = vc.Close()
+// 				if err != nil && err != io.EOF {
+// 					fmt.Printf("\nERROR LEAVING VC: %+v\n", err)
+// 				}
+// 				fmt.Println("\nLEAVING VOICE CHANNEL")
+// 				return
+// 			case <-ticker.C:
+// 				nextFrame, err := encodeSess.OpusFrame()
+// 				if err != nil && err != io.EOF {
+// 					fmt.Printf("\nERROR PLAYING DCA: %+v\n", err)
+// 				}
+// 				if err == io.EOF {
+// 					fmt.Println("\nPLAYBACK FINISHED")
+// 					eofChannel <- true
+// 				}
+// 				err = vc.SendOpusFrame(nextFrame)
+// 			}
+// 		}
+// 	}()
 
-	go func() {
-		for {
-			time.Sleep(1 * time.Second)
-			select {
-			case <-eofChannel:
-				done <- true
-				return
-			}
-		}
-	}()
-}
+// 	go func() {
+// 		for {
+// 			time.Sleep(1 * time.Second)
+// 			select {
+// 			case <-eofChannel:
+// 				done <- true
+// 				return
+// 			}
+// 		}
+// 	}()
+// }
