+++ discord/message.go
@@ -1,6 +1,7 @@
 package discord
 
 import (
+	"fmt"
 	"strings"
 	"time"
 
@@ -34,4 +35,16 @@ func (mec *MessageEventClient) FilterNonModLinks() error {
 		break
 	}
 	return nil
+}
+
+func deleteMessage(resp *disgord.Message, sleep time.Duration, client disgordiface.DisgordClientAPI) {
+	time.Sleep(sleep)
+
+	fmt.Printf("\nDeleting message '%+v' by user %+v \n", resp.Content, resp.Author.Username)
+
+	channel := client.Channel(resp.ChannelID)
+	msgQueryBuilder := channel.Message(resp.ID)
+
+	msgQueryBuilder.Delete()
+
 }
